pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["Krigor Rosa <krigorsilva13@gmail.com>"]
readme = "README.md"
packages = [{include = "discord_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.1"
python-dotenv = "^1.0.0"
openai = "^0.27.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
